tidying, reduce amount of uploaded files, delete assets from draft
Reduce amount of uploaded files:
GitHub has a limit of 1k assets per Release, so it's best to reduce the number of files uploaded.
- Don't upload javadoc (it's not necessary, it's never used by consumers, and source-jar is better).
- Don't upload sha512.

Delete assets from draft:
If a draft already has attached assets, delete them.

+++ modules/lib-github-rest-client/src/main/kotlin/endpoints/repos/DeleteReleaseAsset.kt
@@ -8,6 +8,6 @@ data object DeleteReleaseAsset {
   class Route internal constructor(
     val owner: String,
     val repo: String,
-    val assetId: Int,
+    val assetId: Long,
   )
 }
